pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llama-memory"
version = "1.0.0"
description = "Local vector memory for AI assistants using llama.cpp and sqlite-vec"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "InquiringMinds-AI", email = "support@inquiringminds.ai"}
]
keywords = [
    "llama",
    "memory",
    "vector",
    "embeddings",
    "mcp",
    "claude",
    "ai",
    "local",
    "privacy",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Operating System :: Android",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Database",
]
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
]

[project.urls]
Homepage = "https://github.com/InquiringMinds-AI/llama-memory"
Repository = "https://github.com/InquiringMinds-AI/llama-memory"
Issues = "https://github.com/InquiringMinds-AI/llama-memory/issues"

[project.scripts]
llama-memory = "llama_memory.cli:main"

[tool.setuptools.packages.find]
include = ["llama_memory*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
